Enhance frontend UI
