+++ frontend/src/App.tsx
@@ -1,19 +1,7 @@
-import React, { useEffect, useState, useRef } from 'react';
-
-// Chart.js is loaded globally via a script tag in index.html
-declare const Chart: any;
-
-interface Lead {
-  id: number;
-  name: string;
-  company: string;
-  industry: string;
-  size: number;
-  source: string;
-  created_at: string;
-}
-
-type View = 'table' | 'chart';
+import React, { useEffect, useState } from 'react';
+import FilterBar, { View } from './components/FilterBar';
+import LeadTable, { Lead } from './components/LeadTable';
+import SourceChart from './components/SourceChart';
 
 const App: React.FC = () => {
   const [leads, setLeads] = useState<Lead[]>([]);
@@ -31,10 +19,6 @@ const App: React.FC = () => {
     }
     return id;
   });
-
-  // Refs for rendering the Chart.js graph
-  const canvasRef = useRef<HTMLCanvasElement | null>(null);
-  const chartRef = useRef<any>(null);
 
   const postEvent = async (action: string, metadata: Record<string, any>) => {
     try {
@@ -76,37 +60,6 @@ const App: React.FC = () => {
     fetchLeads();
   }, [industry, size]);
 
-  // Render the chart whenever the data or view changes
-  useEffect(() => {
-    if (view !== 'chart' || !canvasRef.current) return;
-    const ctx = canvasRef.current.getContext('2d');
-    if (!ctx) return;
-
-    if (chartRef.current) {
-      chartRef.current.destroy();
-    }
-
-    chartRef.current = new Chart(ctx, {
-      type: 'bar',
-      data: {
-        labels: Object.keys(sourceCounts),
-        datasets: [
-          {
-            label: 'Leads',
-            data: Object.values(sourceCounts),
-            backgroundColor: 'rgba(54, 162, 235, 0.6)',
-          },
-        ],
-      },
-      options: {
-        responsive: true,
-        plugins: {
-          legend: { display: false },
-        },
-      },
-    });
-  }, [view, sourceCounts]);
-
   const onIndustryChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
     setIndustry(e.target.value);
     postEvent('industry_filter', { industry: e.target.value });
@@ -129,62 +82,22 @@ const App: React.FC = () => {
   };
 
   return (
-    <div style={{ padding: '1rem' }}>
+    <div className="container">
       <h1>Leads Dashboard</h1>
-      <div style={{ marginBottom: '1rem' }}>
-        <label>
-          Industry:
-          <select value={industry} onChange={onIndustryChange}>
-            <option value="">All</option>
-            <option value="Technology">Technology</option>
-            <option value="Manufacturing">Manufacturing</option>
-            <option value="Healthcare">Healthcare</option>
-            <option value="Finance">Finance</option>
-          </select>
-        </label>
-        <label style={{ marginLeft: '1rem' }}>
-          Min Size: {size}
-          <input
-            type="range"
-            min="0"
-            max="500"
-            value={size}
-            onChange={onSizeChange}
-            style={{ width: '10rem', verticalAlign: 'middle', marginLeft: '0.5rem' }}
-          />
-        </label>
-        <button onClick={onRefresh} style={{ marginLeft: '1rem' }}>Refresh</button>
-        <button onClick={() => onToggleView('table')} style={{ marginLeft: '1rem' }}>Table</button>
-        <button onClick={() => onToggleView('chart')} style={{ marginLeft: '0.5rem' }}>Chart</button>
-      </div>
+      <FilterBar
+        industry={industry}
+        size={size}
+        view={view}
+        onIndustryChange={onIndustryChange}
+        onSizeChange={onSizeChange}
+        onRefresh={onRefresh}
+        onToggleView={onToggleView}
+      />
 
       {view === 'table' ? (
-        <table border={1} cellPadding={4} cellSpacing={0}>
-          <thead>
-            <tr>
-              <th>Name</th>
-              <th>Company</th>
-              <th>Industry</th>
-              <th>Size</th>
-              <th>Source</th>
-              <th>Created</th>
-            </tr>
-          </thead>
-          <tbody>
-            {leads.map((lead) => (
-              <tr key={lead.id}>
-                <td>{lead.name}</td>
-                <td>{lead.company}</td>
-                <td>{lead.industry}</td>
-                <td>{lead.size}</td>
-                <td>{lead.source}</td>
-                <td>{new Date(lead.created_at).toLocaleDateString()}</td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
+        <LeadTable leads={leads} />
       ) : (
-        <canvas ref={canvasRef} />
+        <SourceChart counts={sourceCounts} />
       )}
     </div>
   );
